refactor: move form handlers into reusable files
